@@ -1,11 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Collections.Immutable;
-using System.Linq;
-using System.Text;
+﻿using System.Collections.Immutable;
 using System.Text.Json.Nodes;
 using System.Text.Json.Serialization;
-using System.Threading.Tasks;
 using Backend.Messaging.Chain;
 
 namespace Beey.DataExchangeModel.Messaging;
@@ -21,6 +16,7 @@ public static partial class KnownSubsystems
         [JsonSerializable(typeof(Failed))]
         [JsonSerializable(typeof(Completed))]
         public partial class DiarizationSerializerContext : JsonSerializerContext { };
+
         public static string Name => KnownSubsystemNames.DiarizationSubsystem;
         public sealed record Started(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : StartedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
         public sealed record Progress(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, JsonNode Data) : ProgressMessage(Id, Index, ProjectId, ChainId, Name, Sent, Data);
@@ -35,12 +31,14 @@ public static partial class KnownSubsystems
         [JsonSerializable(typeof(Failed))]
         [JsonSerializable(typeof(Completed))]
         public partial class RawDiarizationSerializerContext : JsonSerializerContext { };
+
         public static string Name => KnownSubsystemNames.RawDiarizationSubsystem;
         public sealed record Started(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : StartedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
         public sealed record Progress(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, JsonNode Data) : ProgressMessage(Id, Index, ProjectId, ChainId, Name, Sent, Data);
         public sealed record Failed(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, string Reason) : FailedMessage(Id, Index, ProjectId, ChainId, Name, Sent, Reason);
         public sealed record Completed(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : CompletedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
     }
+
     public static partial class VoiceprintAggregation
     {
         [JsonSerializable(typeof(Started))]
@@ -48,13 +46,14 @@ public static partial class KnownSubsystems
         [JsonSerializable(typeof(Failed))]
         [JsonSerializable(typeof(Completed))]
         public partial class VoiceprintAggregationSerializerContext : JsonSerializerContext { };
+
         public static string Name => KnownSubsystemNames.VoiceprintAggregation;
         public sealed record Started(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : StartedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
         public sealed record Progress(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, JsonNode Data) : ProgressMessage(Id, Index, ProjectId, ChainId, Name, Sent, Data);
         public sealed record Failed(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, string Reason) : FailedMessage(Id, Index, ProjectId, ChainId, Name, Sent, Reason);
         public sealed record Completed(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : CompletedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
-
     }
+
     public static partial class Upload
     {
         [JsonSerializable(typeof(Started))]
@@ -62,12 +61,14 @@ public static partial class KnownSubsystems
         [JsonSerializable(typeof(Failed))]
         [JsonSerializable(typeof(Completed))]
         public partial class UploadSerializerContext : JsonSerializerContext { };
+
         public static string Name => KnownSubsystemNames.UploadSubsystem;
         public sealed record Started(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : StartedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
         public sealed record Progress(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, JsonNode Data) : ProgressMessage(Id, Index, ProjectId, ChainId, Name, Sent, Data);
         public sealed record Failed(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, string Reason) : FailedMessage(Id, Index, ProjectId, ChainId, Name, Sent, Reason);
         public sealed record Completed(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : CompletedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
     }
+
     public static partial class TranscriptionTracking
     {
         [JsonSerializable(typeof(Started))]
@@ -75,12 +76,14 @@ public static partial class KnownSubsystems
         [JsonSerializable(typeof(Failed))]
         [JsonSerializable(typeof(Completed))]
         public partial class TranscriptionTrackingSerializerContext : JsonSerializerContext { };
+
         public static string Name => KnownSubsystemNames.TranscriptionTracking;
         public sealed record Started(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : StartedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
         public sealed record Progress(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, JsonNode Data) : ProgressMessage(Id, Index, ProjectId, ChainId, Name, Sent, Data);
         public sealed record Failed(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, string Reason) : FailedMessage(Id, Index, ProjectId, ChainId, Name, Sent, Reason);
         public sealed record Completed(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : CompletedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
     }
+
     public static partial class TranscriptionTimeLogging
     {
         [JsonSerializable(typeof(Started))]
@@ -88,12 +91,14 @@ public static partial class KnownSubsystems
         [JsonSerializable(typeof(Failed))]
         [JsonSerializable(typeof(Completed))]
         public partial class TranscriptionTimeLoggingSerializerContext : JsonSerializerContext { };
+
         public static string Name => KnownSubsystemNames.TranscriptionTimeLogging;
         public sealed record Started(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : StartedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
         public sealed record Progress(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, JsonNode Data) : ProgressMessage(Id, Index, ProjectId, ChainId, Name, Sent, Data);
         public sealed record Failed(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, string Reason) : FailedMessage(Id, Index, ProjectId, ChainId, Name, Sent, Reason);
         public sealed record Completed(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : CompletedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
     }
+
     public static partial class TranscriptionCreation
     {
         [JsonSerializable(typeof(Started))]
@@ -101,12 +106,14 @@ public static partial class KnownSubsystems
         [JsonSerializable(typeof(Failed))]
         [JsonSerializable(typeof(Completed))]
         public partial class TranscriptionCreationSerializerContext : JsonSerializerContext { };
+
         public static string Name => KnownSubsystemNames.TranscriptionCreation;
         public sealed record Started(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : StartedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
         public sealed record Progress(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, JsonNode Data) : ProgressMessage(Id, Index, ProjectId, ChainId, Name, Sent, Data);
         public sealed record Failed(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, string Reason) : FailedMessage(Id, Index, ProjectId, ChainId, Name, Sent, Reason);
         public sealed record Completed(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : CompletedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
     }
+
     public static partial class MediaFilePackaging
     {
         [JsonSerializable(typeof(Started))]
@@ -114,12 +121,14 @@ public static partial class KnownSubsystems
         [JsonSerializable(typeof(Failed))]
         [JsonSerializable(typeof(Completed))]
         public partial class MediaFilePackagingSerializerContext : JsonSerializerContext { };
+
         public static string Name => KnownSubsystemNames.MediaFilePackagingSubsystem;
         public sealed record Started(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : StartedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
         public sealed record Progress(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, JsonNode Data) : ProgressMessage(Id, Index, ProjectId, ChainId, Name, Sent, Data);
         public sealed record Failed(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, string Reason) : FailedMessage(Id, Index, ProjectId, ChainId, Name, Sent, Reason);
         public sealed record Completed(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : CompletedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
     }
+
     public static partial class CreditReservation
     {
         [JsonSerializable(typeof(Started))]
@@ -127,12 +136,14 @@ public static partial class KnownSubsystems
         [JsonSerializable(typeof(Failed))]
         [JsonSerializable(typeof(Completed))]
         public partial class CreditReservationSerializerContext : JsonSerializerContext { };
+
         public static string Name => KnownSubsystemNames.CreditReservation;
         public sealed record Started(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : StartedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
         public sealed record Progress(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, JsonNode Data) : ProgressMessage(Id, Index, ProjectId, ChainId, Name, Sent, Data);
         public sealed record Failed(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, string Reason) : FailedMessage(Id, Index, ProjectId, ChainId, Name, Sent, Reason);
         public sealed record Completed(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : CompletedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
     }
+
     public static partial class MediaFileIndexing
     {
         [JsonSerializable(typeof(Started))]
@@ -140,12 +151,14 @@ public static partial class KnownSubsystems
         [JsonSerializable(typeof(Failed))]
         [JsonSerializable(typeof(Completed))]
         public partial class MediaFileIndexingSerializerContext : JsonSerializerContext { };
+
         public static string Name => KnownSubsystemNames.MediaFileIndexingSubsystem;
         public sealed record Started(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : StartedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
         public sealed record Progress(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, JsonNode Data) : ProgressMessage(Id, Index, ProjectId, ChainId, Name, Sent, Data);
         public sealed record Failed(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, string Reason) : FailedMessage(Id, Index, ProjectId, ChainId, Name, Sent, Reason);
         public sealed record Completed(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : CompletedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
     }
+
     public static partial class MediaIdentification
     {
         [JsonSerializable(typeof(Started))]
@@ -153,12 +166,14 @@ public static partial class KnownSubsystems
         [JsonSerializable(typeof(Failed))]
         [JsonSerializable(typeof(Completed))]
         public partial class MediaIdentificationSerializerContext : JsonSerializerContext { };
+
         public static string Name => KnownSubsystemNames.MediaIdentificationSubsystem;
         public sealed record Started(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : StartedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
         public sealed record Progress(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, JsonNode Data) : ProgressMessage(Id, Index, ProjectId, ChainId, Name, Sent, Data);
         public sealed record Failed(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, string Reason) : FailedMessage(Id, Index, ProjectId, ChainId, Name, Sent, Reason);
         public sealed record Completed(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : CompletedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
     }
+
     public static partial class ProjectStatusMonitor
     {
         [JsonSerializable(typeof(Started))]
@@ -166,12 +181,14 @@ public static partial class KnownSubsystems
         [JsonSerializable(typeof(Failed))]
         [JsonSerializable(typeof(Completed))]
         public partial class ProjectStatusMonitorSerializerContext : JsonSerializerContext { };
+
         public static string Name => KnownSubsystemNames.ProjectStatusMonitor;
         public sealed record Started(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : StartedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
         public sealed record Progress(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, JsonNode Data) : ProgressMessage(Id, Index, ProjectId, ChainId, Name, Sent, Data);
         public sealed record Failed(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, string Reason) : FailedMessage(Id, Index, ProjectId, ChainId, Name, Sent, Reason);
         public sealed record Completed(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : CompletedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
     }
+
     public static partial class Recognition
     {
         [JsonSerializable(typeof(Started))]
@@ -179,6 +196,7 @@ public static partial class KnownSubsystems
         [JsonSerializable(typeof(Failed))]
         [JsonSerializable(typeof(Completed))]
         public partial class RecognitionSerializerContext : JsonSerializerContext { };
+
         public static string Name => KnownSubsystemNames.RecognitionSubsystem;
         public sealed record Started(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : StartedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
         public sealed record Progress(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, JsonNode Data) : ProgressMessage(Id, Index, ProjectId, ChainId, Name, Sent, Data);
@@ -193,12 +211,14 @@ public static partial class KnownSubsystems
         [JsonSerializable(typeof(Failed))]
         [JsonSerializable(typeof(Completed))]
         public partial class RawRecognitionSerializerContext : JsonSerializerContext { };
+
         public static string Name => KnownSubsystemNames.RawRecognitionSubsystem;
         public sealed record Started(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : StartedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
         public sealed record Progress(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, JsonNode Data) : ProgressMessage(Id, Index, ProjectId, ChainId, Name, Sent, Data);
         public sealed record Failed(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, string Reason) : FailedMessage(Id, Index, ProjectId, ChainId, Name, Sent, Reason);
         public sealed record Completed(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : CompletedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
     }
+
     public static partial class SpeakerId
     {
         [JsonSerializable(typeof(Started))]
@@ -206,12 +226,14 @@ public static partial class KnownSubsystems
         [JsonSerializable(typeof(Failed))]
         [JsonSerializable(typeof(Completed))]
         public partial class SpeakerIdentificationSerializerContext : JsonSerializerContext { };
+
         public static string Name => KnownSubsystemNames.SpeakerIdentificationSubsystem;
         public sealed record Started(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : StartedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
         public sealed record Progress(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, JsonNode Data) : ProgressMessage(Id, Index, ProjectId, ChainId, Name, Sent, Data);
         public sealed record Failed(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, string Reason) : FailedMessage(Id, Index, ProjectId, ChainId, Name, Sent, Reason);
         public sealed record Completed(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : CompletedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
     }
+
     public static partial class Spp
     {
         [JsonSerializable(typeof(Started))]
@@ -219,12 +241,14 @@ public static partial class KnownSubsystems
         [JsonSerializable(typeof(Failed))]
         [JsonSerializable(typeof(Completed))]
         public partial class SppSerializerContext : JsonSerializerContext { };
+
         public static string Name => KnownSubsystemNames.SppSubsystem;
         public sealed record Started(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : StartedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
         public sealed record Progress(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, JsonNode Data) : ProgressMessage(Id, Index, ProjectId, ChainId, Name, Sent, Data);
         public sealed record Failed(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, string Reason) : FailedMessage(Id, Index, ProjectId, ChainId, Name, Sent, Reason);
         public sealed record Completed(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : CompletedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
     }
+
     public static partial class TranscodingVideo
     {
         [JsonSerializable(typeof(Started))]
@@ -232,12 +256,14 @@ public static partial class KnownSubsystems
         [JsonSerializable(typeof(Failed))]
         [JsonSerializable(typeof(Completed))]
         public partial class TranscodingVideoSerializerContext : JsonSerializerContext { };
+
         public static string Name => KnownSubsystemNames.TranscodingVideoSubsystem;
         public sealed record Started(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : StartedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
         public sealed record Progress(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, JsonNode Data) : ProgressMessage(Id, Index, ProjectId, ChainId, Name, Sent, Data);
         public sealed record Failed(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, string Reason) : FailedMessage(Id, Index, ProjectId, ChainId, Name, Sent, Reason);
         public sealed record Completed(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : CompletedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
     }
+
     public static partial class TranscodingAudio
     {
         [JsonSerializable(typeof(Started))]
@@ -245,6 +271,7 @@ public static partial class KnownSubsystems
         [JsonSerializable(typeof(Failed))]
         [JsonSerializable(typeof(Completed))]
         public partial class TranscodingAudioSerializerContext : JsonSerializerContext { };
+
         public static string Name => KnownSubsystemNames.TranscodingAudioSubsystem;
         public sealed record Started(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : StartedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
         public sealed record Progress(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, JsonNode Data) : ProgressMessage(Id, Index, ProjectId, ChainId, Name, Sent, Data);
@@ -254,7 +281,6 @@ public static partial class KnownSubsystems
 
     public static partial class TranscriptionQueueTracking
     {
-
         [JsonSerializable(typeof(Started))]
         [JsonSerializable(typeof(Completed))]
         public partial class TranscriptionQueueTrackingSerializerContext : JsonSerializerContext { };
@@ -266,7 +292,6 @@ public static partial class KnownSubsystems
 
     public static partial class LowQualityAudio
     {
-
         [JsonSerializable(typeof(Started))]
         [JsonSerializable(typeof(Completed))]
         [JsonSerializable(typeof(Failed))]
@@ -277,9 +302,9 @@ public static partial class KnownSubsystems
         public sealed record Completed(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : CompletedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
         public sealed record Failed(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, string Reason) : FailedMessage(Id, Index, ProjectId, ChainId, Name, Sent, Reason);
     }
+
     public static partial class SceneDetection
     {
-
         [JsonSerializable(typeof(Started))]
         [JsonSerializable(typeof(Completed))]
         [JsonSerializable(typeof(Failed))]
@@ -317,6 +342,7 @@ public static partial class KnownSubsystems
         [JsonSerializable(typeof(Failed))]
         [JsonSerializable(typeof(Completed))]
         public partial class TranscriptionStreamingSerializerContext : JsonSerializerContext { };
+
         public static string Name => KnownSubsystemNames.TranscriptionStreamingSubsystem;
         public sealed record Started(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : StartedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
         public sealed record Progress(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, JsonNode Data) : ProgressMessage(Id, Index, ProjectId, ChainId, Name, Sent, Data);
@@ -331,6 +357,7 @@ public static partial class KnownSubsystems
         [JsonSerializable(typeof(Failed))]
         [JsonSerializable(typeof(Completed))]
         public partial class LiveSubtitlesStreamingSerializerContext : JsonSerializerContext { };
+
         public static string Name => KnownSubsystemNames.LiveSubtitlesStreamingSubsystem;
         public sealed record Started(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent) : StartedMessage(Id, Index, ProjectId, ChainId, Name, Sent);
         public sealed record Progress(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, DateTimeOffset Sent, JsonNode Data) : ProgressMessage(Id, Index, ProjectId, ChainId, Name, Sent, Data);
@@ -342,6 +369,7 @@ public static partial class KnownSubsystems
     {
         [JsonSerializable(typeof(Progress))]
         public partial class ProjectUpdatesSerializerContext : JsonSerializerContext { };
+
         public static string Name => KnownSubsystemNames.ProjectUpdates;
         public sealed record Progress(int? ProjectId, int? ChainId, JsonNode Data)
             : ProgressMessage(-1, new SubsystemNodeIndex(), ProjectId, ChainId, Name, DateTimeOffset.Now, Data);
