@@ -1,5 +1,5 @@
-﻿using System;
-using System.Collections.Immutable;
+﻿using System.Collections.Immutable;
+using System.Text.Json.Serialization;
 
 namespace Beey.DataExchangeModel.Messaging;
 
@@ -15,5 +15,6 @@ namespace Beey.DataExchangeModel.Messaging;
 public sealed record PanicMessage(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, string Subsystem, DateTimeOffset Sent, string Reason)
     : Message(Id, Index, ProjectId, ChainId, Subsystem, Sent)
 {
+    [JsonPropertyOrder(int.MinValue)]//always must be second for deserialization to work
     public override MessageType Type => MessageType.Panic;
 }
