@@ -1,10 +1,11 @@
-﻿using System;
-using System.Collections.Immutable;
+﻿using System.Collections.Immutable;
 using System.Text.Json.Nodes;
+using System.Text.Json.Serialization;
 
 namespace Beey.DataExchangeModel.Messaging;
 
 public record ProgressMessage(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, string Subsystem, DateTimeOffset Sent, JsonNode Data) : Message(Id, Index, ProjectId, ChainId, Subsystem, Sent)
 {
+    [JsonPropertyOrder(int.MinValue)]//always must be second for deserialization to work
     public override MessageType Type => MessageType.Progress;
 }
