@@ -1,14 +1,13 @@
-﻿using Beey.DataExchangeModel.Messaging.Subsystems;
-using Microsoft.Extensions.Configuration;
-using System;
-using System.Collections.Immutable;
+﻿using System.Collections.Immutable;
 using System.Text.Json.Nodes;
+using System.Text.Json.Serialization;
 
 namespace Beey.DataExchangeModel.Messaging;
 
 public record StartedMessage(int Id, ImmutableArray<int> Index, int? ProjectId, int? ChainId, string Subsystem, DateTimeOffset Sent)
     : Message(Id, Index, ProjectId, ChainId, Subsystem, Sent)
 {
+    [JsonPropertyOrder(int.MinValue)]//always must be second for deserialization to work
     public override MessageType Type => MessageType.Started;
 
     [Obsolete("just for backwards compatibility with deserializers")]
